+++ pytorch_lightning/tuner/batch_size_scaling.py
@@ -166,6 +166,7 @@ def _run_power_scaling(
         if changed:
             # Force the train dataloader to reset as the batch size has changed
             trainer.reset_train_dataloader(model)
+            trainer.reset_val_dataloader(model)
         else:
             break
     return new_size